Refactor renderer surface caching
