+++ src/heart/utilities/reactivex_streams.py
@@ -6,6 +6,7 @@ from heart.utilities.reactivex.instrumentation import \
     _STATS_SCHEDULER  # noqa: F401
 from heart.utilities.reactivex.settings import \
     StreamShareSettings  # noqa: F401
-from heart.utilities.reactivex.sharing import (_GRACE_SCHEDULER,  # noqa: F401
-                                               _REPLAY_SCHEDULER, share_stream)
+from heart.utilities.reactivex.sharing import _GRACE_SCHEDULER  # noqa: F401
+from heart.utilities.reactivex.sharing import _REPLAY_SCHEDULER  # noqa: F401
+from heart.utilities.reactivex.sharing import share_stream  # noqa: F401
 from heart.utilities.reactivex.types import ConnectableStream  # noqa: F401
